[INLONG-6663][Manager] Move MQType from inlong-manager to inlong-common (#6664)
