+++ inlong-common/src/main/java/org/apache/inlong/common/constant/MQType.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package org.apache.inlong.manager.common.consts;
+package org.apache.inlong.common.constant;
 
 /**
  * Constants of MQ type.
