+++ inlong-manager/manager-client/src/test/java/org/apache/inlong/manager/client/api/inner/InlongConsumeClientTest.java
@@ -19,7 +19,7 @@ package org.apache.inlong.manager.client.api.inner;
 
 import com.google.common.collect.Lists;
 import org.apache.inlong.manager.client.api.inner.client.InlongConsumeClient;
-import org.apache.inlong.manager.common.consts.MQType;
+import org.apache.inlong.common.constant.MQType;
 import org.apache.inlong.manager.common.util.JsonUtils;
 import org.apache.inlong.manager.pojo.common.PageResult;
 import org.apache.inlong.manager.pojo.common.Response;
