+++ inlong-manager/manager-pojo/src/main/java/org/apache/inlong/manager/pojo/group/none/InlongNoneMqRequest.java
@@ -21,7 +21,7 @@ import io.swagger.annotations.ApiModel;
 import lombok.Data;
 import lombok.EqualsAndHashCode;
 import lombok.ToString;
-import org.apache.inlong.manager.common.consts.MQType;
+import org.apache.inlong.common.constant.MQType;
 import org.apache.inlong.manager.pojo.group.InlongGroupRequest;
 import org.apache.inlong.manager.common.util.JsonTypeDefine;
